Generate SOK from current Gatherer and patch regressions.
We are replacing some translations here but note the translations we are
replacing are the WRONG translations. Previously the front side of these cards
had the translated name of the flip side. We preserve the translations of the
flip side with a patch, the front side becomes English as we don't have
translations for those.