Patch Gatherer regressions, remove bad translations and provide text for Homura's Essence in SOK (#503)
* Generate SOK from current Gatherer and patch regressions.

We are replacing some translations here but note the translations we are
replacing are the WRONG translations. Previously the front side of these cards
had the translated name of the flip side. We preserve the translations of the
flip side with a patch, the front side becomes English as we don't have
translations for those.

* Patch missing text in Gatherer in SOK config file.

Fixes #498.

* Update changelog.json for latest changes.